+++ java/rocksjni/concurrent_task_limiter.cc
@@ -3,16 +3,16 @@
 #include <memory>
 #include <string>
 
-#include "include/org_rocksdb_limiter_ConcurrentTaskLimiterImpl.h"
+#include "include/org_rocksdb_ConcurrentTaskLimiterImpl.h"
 #include "rocksdb/concurrent_task_limiter.h"
 #include "rocksjni/portal.h"
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    newConcurrentTaskLimiterImpl0
  * Signature: (Ljava/lang/String;I)J
  */
-jlong Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_newConcurrentTaskLimiterImpl0(
+jlong Java_org_rocksdb_ConcurrentTaskLimiterImpl_newConcurrentTaskLimiterImpl0(
     JNIEnv *env, jclass, jstring jname, jint limit) {
   jboolean has_exception;
   std::string name = ROCKSDB_NAMESPACE::JniUtil::copyStdString(env, jname, &has_exception);
@@ -27,11 +27,11 @@ jlong Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_newConcurrentTaskLimite
 }
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    name
  * Signature: (J)Ljava/lang/String;
  */
-jstring Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_name
+jstring Java_org_rocksdb_ConcurrentTaskLimiterImpl_name
   (JNIEnv *env, jclass, jlong handle) {
   const auto& limiter =
       *reinterpret_cast<std::shared_ptr<ROCKSDB_NAMESPACE::ConcurrentTaskLimiter>*>(handle);
@@ -39,11 +39,11 @@ jstring Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_name
 }
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    setMaxOutstandingTask
  * Signature: (JI)V
  */
-void Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_setMaxOutstandingTask
+void Java_org_rocksdb_ConcurrentTaskLimiterImpl_setMaxOutstandingTask
   (JNIEnv *, jclass, jlong handle, jint max_outstanding_task) {
   const auto& limiter =
       *reinterpret_cast<std::shared_ptr<ROCKSDB_NAMESPACE::ConcurrentTaskLimiter>*>(handle);
@@ -51,11 +51,11 @@ void Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_setMaxOutstandingTask
 }
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    resetMaxOutstandingTask
  * Signature: (J)V
  */
-void Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_resetMaxOutstandingTask
+void Java_org_rocksdb_ConcurrentTaskLimiterImpl_resetMaxOutstandingTask
   (JNIEnv *, jclass, jlong handle) {
   const auto& limiter =
       *reinterpret_cast<std::shared_ptr<ROCKSDB_NAMESPACE::ConcurrentTaskLimiter>*>(handle);
@@ -63,11 +63,11 @@ void Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_resetMaxOutstandingTask
 }
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    outstandingTask
  * Signature: (J)I
  */
-jint Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_outstandingTask
+jint Java_org_rocksdb_ConcurrentTaskLimiterImpl_outstandingTask
   (JNIEnv *, jclass, jlong handle) {
   const auto& limiter =
       *reinterpret_cast<std::shared_ptr<ROCKSDB_NAMESPACE::ConcurrentTaskLimiter>*>(handle);
@@ -75,11 +75,11 @@ jint Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_outstandingTask
 }
 
 /*
- * Class:     org_rocksdb_limiter_ConcurrentTaskLimiterImpl
+ * Class:     org_rocksdb_ConcurrentTaskLimiterImpl
  * Method:    disposeInternal
  * Signature: (J)V
  */
-void Java_org_rocksdb_limiter_ConcurrentTaskLimiterImpl_disposeInternal(
+void Java_org_rocksdb_ConcurrentTaskLimiterImpl_disposeInternal(
     JNIEnv *, jobject, jlong jhandle) {
   auto* ptr = reinterpret_cast<
       std::shared_ptr<ROCKSDB_NAMESPACE::ConcurrentTaskLimiter>*>(jhandle);
