+++ java/src/main/java/org/rocksdb/ConcurrentTaskLimiter.java
@@ -1,6 +1,4 @@
-package org.rocksdb.limiter;
-
-import org.rocksdb.RocksObject;
+package org.rocksdb;
 
 public abstract class ConcurrentTaskLimiter extends RocksObject {
   protected ConcurrentTaskLimiter(final long nativeHandle) {
